dns: Move CheckCRCLocalDNSReachableFromHost from network to dns package
The 'dns' package already contains the other CheckCRC*Reachable*
functions, so it makes sense to also have CheckCRCLocalDNSReachableFromHost
This remove the cyclic dependency introduced in the previous commit.
