+++ pkg/crc/services/dns/dns.go
@@ -3,11 +3,14 @@ package dns
 import (
 	"context"
 	"fmt"
+	"net"
+	"runtime"
 	"time"
 
 	"github.com/crc-org/crc/pkg/crc/adminhelper"
 	"github.com/crc-org/crc/pkg/crc/constants"
 	"github.com/crc-org/crc/pkg/crc/errors"
+	"github.com/crc-org/crc/pkg/crc/logging"
 	"github.com/crc-org/crc/pkg/crc/network"
 	"github.com/crc-org/crc/pkg/crc/services"
 	"github.com/crc-org/crc/pkg/crc/systemd"
@@ -140,6 +143,51 @@ func CheckCRCPublicDNSReachable(serviceConfig services.ServicePostStartConfig) (
 	}
 	stdout, _, err := serviceConfig.SSHRunner.Run("curl", curlArgs...)
 	return stdout, err
+}
+
+func CheckCRCLocalDNSReachableFromHost(apiHostname, appsHostname, appsDomain, expectedIP string) error {
+	ip, err := net.LookupIP(apiHostname)
+	if err != nil {
+		return err
+	}
+	logging.Debugf("%s resolved to %s", apiHostname, ip)
+	if !matchIP(ip, expectedIP) {
+		logging.Warnf("%s resolved to %s but %s was expected", apiHostname, ip, expectedIP)
+		return fmt.Errorf("Invalid IP for %s", apiHostname)
+	}
+
+	if runtime.GOOS != "darwin" {
+		/* This check will fail with !CGO_ENABLED builds on darwin as
+		 * in this case, /etc/resolver/ will not be used, so we won't
+		 * have wildcard DNS for our domains
+		 */
+		ip, err = net.LookupIP(appsHostname)
+		if err != nil {
+			// Right now admin helper fallback is not implemented on windows so
+			// this check should still return an error.
+			if runtime.GOOS == "windows" {
+				return err
+			}
+			logging.Warnf("Wildcard DNS resolution for %s does not appear to be working", appsDomain)
+			return nil
+		}
+		logging.Debugf("%s resolved to %s", appsHostname, ip)
+		if !matchIP(ip, expectedIP) {
+			logging.Warnf("%s resolved to %s but %s was expected", appsHostname, ip, expectedIP)
+			return fmt.Errorf("Invalid IP for %s", appsHostname)
+		}
+	}
+	return nil
+}
+
+func matchIP(ips []net.IP, expectedIP string) bool {
+	for _, ip := range ips {
+		if ip.String() == expectedIP {
+			return true
+		}
+	}
+
+	return false
 }
 
 func addOpenShiftHosts(serviceConfig services.ServicePostStartConfig) error {
